refactor: cleanup unnecessary table tool utils
